Combine filtered and regular data pulls
