--- scripts/google_analytics/config.json
@@ -1,10 +1,0 @@
-{
-  "core": "111433053",
-  "hca": "123860701",
-  "gibct": "111454100",
-  "vec": "84936422",
-  "facility": "130690170",
-  "claims": "132929836",
-  "rx": "132496709",
-  "messaging": "132909478"
-}
