+++ scripts/google_analytics/update.sh
@@ -19,6 +19,7 @@ bin/pip3 install -r requirements.txt
 export GA_SERVICEACCOUNT=serviceaccount.p12
 
 bin/python3 update_data.py
+bin/python3 update_clicks.py
 
 mv *.csv $DATA_DIR
